# swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Dev Mitra API",
    "version": "1.0.0",
    "description": "Dev Mitra Backend API Documentation"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "summary": "Login with phone and OTP",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "phone",
                  "otp"
                ],
                "properties": {
                  "phone": {
                    "type": "string",
                    "example": "9876543210"
                  },
                  "otp": {
                    "type": "string",
                    "example": "123456"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          }
        }
      }
    },
...
    "/api/upload/profile": {
      "post": {
        "summary": "Upload profile image (Customer or Vendor)",
        "tags": [
          "Upload"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "image"
                ],
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Image uploaded successfully"
          },
          "400": {
            "description": "No file uploaded"
          }
        }
      }
    },
    "/api/upload/home-card": {
      "post": {
        "summary": "Upload home card image (Admin only)",
        "tags": [
          "Upload"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "image"
                ],
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Home card image uploaded successfully"
          },
          "400": {
            "description": "No file uploaded"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
    "/api/upload/puja": {
      "post": {
        "summary": "Upload puja image (Admin only)",
        "tags": [
          "Upload"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "image"
                ],
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Puja image uploaded successfully"
          },
          "400": {
            "description": "No file uploaded"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
    "/api/upload/icon": {
      "post": {
        "summary": "Upload icon image (Admin only)",
        "tags": [
          "Upload"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "image"
                ],
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Icon uploaded successfully"
          },
          "400": {
            "description": "No file uploaded"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
...
  },
...
}